修复status指标查询逻辑,改为时间控制,移除了 max_over_time 包装，直接使用原始查询语句
