+++ config/config.yaml
@@ -1,5 +1,6 @@
 prometheus_url: "http://prometheus.monitoring.svc.cluster.local:9090"
 
+
 metric_types:
   - type: "基础资源使用情况"
     metrics:
@@ -93,7 +94,7 @@ metric_types:
         description: "持久化存储使用率"
         query: >-
           100 * (1 - kubelet_volume_stats_available_bytes / kubelet_volume_stats_capacity_bytes)
-        threshold: 80
+        threshold: 90
         threshold_type: "greater"
         unit: "%"
         labels:
